end part 2

+++ stattypes.h
@@ -1,3 +1,4 @@
 #pragma once
 #include <cstdint>
-typedef std::uint8_t stattype;
+
+typedef std::uint16_t stattype;
